fix: secure Gemini API key by moving generation to server (#174)
