+++ .env.example
@@ -3,6 +3,4 @@ CLERK_SECRET_KEY="your-CLERK_SECRET_KEY"
 DATABASE_URL = "neonDB datbase url"
 NEXT_PUBLIC_CLERK_SIGN_IN_FALLBACK_REDIRECT_URL=/dashboard
 NEXT_PUBLIC_CLERK_SIGN_UP_FALLBACK_REDIRECT_URL=/dashboard
-
-
-
+GOOGLE_GEMINI_API_KEY="your-GEMINI_API_KEY"
