+++ app/actions/aiActions.ts
@@ -1,0 +1,39 @@
+"use server";
+
+const {
+  GoogleGenerativeAI,
+} = require("@google/generative-ai");
+
+export async function generateAiContent(prompt: string) {
+  try {
+    const apiKey = process.env.GOOGLE_GEMINI_API_KEY;
+    if (!apiKey) {
+      throw new Error("API key is not configured");
+    }
+
+    const genAI = new GoogleGenerativeAI(apiKey);
+    
+    const model = genAI.getGenerativeModel({
+      model: "gemini-1.5-flash",
+    });
+    
+    const generationConfig = {
+      temperature: 1,
+      topP: 0.95,
+      topK: 64,
+      maxOutputTokens: 8192,
+      responseMimeType: "text/plain",
+    };
+    
+    const chatSession = model.startChat({
+      generationConfig,
+      history: [],
+    });
+
+    const result = await chatSession.sendMessage(prompt);
+    return result.response.text();
+  } catch (error) {
+    console.error("Error in AI generation:", error);
+    throw new Error("Failed to generate AI content");
+  }
+}
